feat: Create a helper functions to help with making the cpu smarter, fr?
feat: create a function called get adjacency slot that recieves a
coordinate that returns a filtered array items of the possible adjacent
of this coordinate.

feat: create a function called get separated coordinates that returns
letter and a number. this is so i can see the pattern of which the
next move should be like say for example if 1a 2a 3a predict the next
move wil be 4a to pick from the adjacent slot of previous 3a 2a and 1a

todo: make the ai smarter i don't know if it's final but it takes time
fr to create one hayst but it's worth it thank GOD.