feat: Create a function that acts like set unvisited but adjacent
problem: when a player ship got hit you want to push adjacent pair
coordinates inside the array adjacent slots. when the function recurse
it will see to it that adjacent slots is not empty and set unvisited
adjacent slot return value will be use as a coordinate for attacking
player ship close to that ship.

todo: it is not very intelligent at the moment but it does the job